requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# LLM clients
openai>=1.12.0
google-generativeai>=0.5.0
anthropic>=0.25.0
together>=0.2.10

# Data processing
scikit-learn>=1.3.0
openpyxl>=3.1.0  # For Excel support

# Utilities
tenacity>=8.2.0